requirements.txt
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
matplotlib>=3.8
PyYAML>=6.0
